+++ Blueprint.Api/Controllers/TeamController.cs
@@ -6,10 +6,11 @@ using System.Collections.Generic;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Blueprint.Api.Infrastructure.Extensions;
 using Blueprint.Api.Infrastructure.Exceptions;
+using Blueprint.Api.Infrastructure.Authorization;
+using Blueprint.Api.Data.Enumerations;
 using Blueprint.Api.Services;
 using Blueprint.Api.ViewModels;
 using Swashbuckle.AspNetCore.Annotations;
@@ -19,9 +20,9 @@ namespace Blueprint.Api.Controllers
     public class TeamController : BaseController
     {
         private readonly ITeamService _teamService;
-        private readonly IAuthorizationService _authorizationService;
+        private readonly IBlueprintAuthorizationService _authorizationService;
 
-        public TeamController(ITeamService teamService, IAuthorizationService authorizationService)
+        public TeamController(ITeamService teamService, IBlueprintAuthorizationService authorizationService)
         {
             _teamService = teamService;
             _authorizationService = authorizationService;
@@ -74,7 +75,8 @@ namespace Blueprint.Api.Controllers
         [SwaggerOperation(OperationId = "getTeamsByMsel")]
         public async Task<IActionResult> GetByMsel([FromRoute] Guid mselId, CancellationToken ct)
         {
-            var list = await _teamService.GetByMselAsync(mselId, ct);
+            var hasSystemPermission = await _authorizationService.AuthorizeAsync([SystemPermission.ViewMsels], ct);
+            var list = await _teamService.GetByMselAsync(mselId, hasSystemPermission, ct);
             return Ok(list);
         }
 
@@ -92,6 +94,9 @@ namespace Blueprint.Api.Controllers
         [SwaggerOperation(OperationId = "getTeamsByUser")]
         public async Task<IActionResult> GetByUser([FromRoute] Guid userId, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.ManageUsers], ct))
+                throw new ForbiddenException();
+
             var list = await _teamService.GetByUserAsync(userId, ct);
             return Ok(list);
         }
@@ -112,6 +117,9 @@ namespace Blueprint.Api.Controllers
         [SwaggerOperation(OperationId = "getTeam")]
         public async Task<IActionResult> Get(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.ViewMsels], ct))
+                throw new ForbiddenException();
+
             var team = await _teamService.GetAsync(id, ct);
 
             if (team == null)
@@ -135,8 +143,9 @@ namespace Blueprint.Api.Controllers
         [SwaggerOperation(OperationId = "createTeam")]
         public async Task<IActionResult> Create([FromBody] Team team, CancellationToken ct)
         {
+            var hasSystemPermission = await _authorizationService.AuthorizeAsync([SystemPermission.EditMsels], ct);
             team.CreatedBy = User.GetId();
-            var createdTeam = await _teamService.CreateAsync(team, ct);
+            var createdTeam = await _teamService.CreateAsync(team, hasSystemPermission, ct);
             return CreatedAtAction(nameof(this.Get), new { id = createdTeam.Id }, createdTeam);
         }
 
@@ -156,7 +165,8 @@ namespace Blueprint.Api.Controllers
         [SwaggerOperation(OperationId = "createTeamFromUnit")]
         public async Task<IActionResult> CreateTeamFromUnit(Guid mselId, Guid unitId, CancellationToken ct)
         {
-            var createdTeam = await _teamService.CreateFromUnitAsync(unitId, mselId, ct);
+            var hasSystemPermission = await _authorizationService.AuthorizeAsync([SystemPermission.EditMsels], ct);
+            var createdTeam = await _teamService.CreateFromUnitAsync(unitId, mselId, hasSystemPermission, ct);
             return CreatedAtAction(nameof(this.Get), new { id = createdTeam.Id }, createdTeam);
         }
 
@@ -176,6 +186,9 @@ namespace Blueprint.Api.Controllers
         [SwaggerOperation(OperationId = "updateTeam")]
         public async Task<IActionResult> Update([FromRoute] Guid id, [FromBody] Team team, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.ManageUsers], ct))
+                throw new ForbiddenException();
+
             team.ModifiedBy = User.GetId();
             var updatedTeam = await _teamService.UpdateAsync(id, team, ct);
             return Ok(updatedTeam);
@@ -196,6 +209,9 @@ namespace Blueprint.Api.Controllers
         [SwaggerOperation(OperationId = "deleteTeam")]
         public async Task<IActionResult> Delete(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.ManageUsers], ct))
+                throw new ForbiddenException();
+
             await _teamService.DeleteAsync(id, ct);
             return NoContent();
         }
